+++ mfa/model_device_authentication_policy_totp.go
@@ -26,6 +26,8 @@ type DeviceAuthenticationPolicyTotp struct {
 	Otp DeviceAuthenticationPolicyTotpOtp `json:"otp"`
 	// Set to `true` if you want to allow users to provide nicknames for devices during pairing.
 	PromptForNicknameOnPairing *bool `json:"promptForNicknameOnPairing,omitempty"`
+	// Object that you can use to provide key:value pairs for `otpauth` URI parameters. For example, if you provide a value for the `issuer` parameter, then authenticators that support that parameter will display the text you specify together with the OTP (in addition to the username). This can help users recognize which application the OTP is for. If you intend on using the same MFA policy for multiple applications, choose a name that reflects the group of applications.
+	UriParameters *map[string]string `json:"uriParameters,omitempty"`
 }
 
 // NewDeviceAuthenticationPolicyTotp instantiates a new DeviceAuthenticationPolicyTotp object
@@ -159,6 +161,38 @@ func (o *DeviceAuthenticationPolicyTotp) SetPromptForNicknameOnPairing(v bool) {
 	o.PromptForNicknameOnPairing = &v
 }
 
+// GetUriParameters returns the UriParameters field value if set, zero value otherwise.
+func (o *DeviceAuthenticationPolicyTotp) GetUriParameters() map[string]string {
+	if o == nil || IsNil(o.UriParameters) {
+		var ret map[string]string
+		return ret
+	}
+	return *o.UriParameters
+}
+
+// GetUriParametersOk returns a tuple with the UriParameters field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *DeviceAuthenticationPolicyTotp) GetUriParametersOk() (*map[string]string, bool) {
+	if o == nil || IsNil(o.UriParameters) {
+		return nil, false
+	}
+	return o.UriParameters, true
+}
+
+// HasUriParameters returns a boolean if a field has been set.
+func (o *DeviceAuthenticationPolicyTotp) HasUriParameters() bool {
+	if o != nil && !IsNil(o.UriParameters) {
+		return true
+	}
+
+	return false
+}
+
+// SetUriParameters gets a reference to the given map[string]string and assigns it to the UriParameters field.
+func (o *DeviceAuthenticationPolicyTotp) SetUriParameters(v map[string]string) {
+	o.UriParameters = &v
+}
+
 func (o DeviceAuthenticationPolicyTotp) MarshalJSON() ([]byte, error) {
 	toSerialize,err := o.ToMap()
 	if err != nil {
@@ -176,6 +210,9 @@ func (o DeviceAuthenticationPolicyTotp) ToMap() (map[string]interface{}, error) 
 	toSerialize["otp"] = o.Otp
 	if !IsNil(o.PromptForNicknameOnPairing) {
 		toSerialize["promptForNicknameOnPairing"] = o.PromptForNicknameOnPairing
+	}
+	if !IsNil(o.UriParameters) {
+		toSerialize["uriParameters"] = o.UriParameters
 	}
 	return toSerialize, nil
 }